Refactored the names of display session view and retired session view and checked the root route, which did not need to be updated
